Refactor network service into to services
